+++ src/ProgramBuilder.ts
@@ -1,6 +1,6 @@
 import * as debounce from 'debounce-promise';
 import * as path from 'path';
-import * as rokuDeploy from 'roku-deploy';
+import { rokuDeploy } from 'roku-deploy';
 import type { BsConfig } from './BsConfig';
 import type { BscFile, BsDiagnostic, FileObj, FileResolver } from './interfaces';
 import { Program } from './Program';
@@ -304,7 +304,7 @@ export class ProgramBuilder {
                 filePath = path.relative(cwd, filePath);
             }
             //load the file text
-            const file = this.program.getFile(srcPath);
+            const file = this.program?.getFile(srcPath);
             //get the file's in-memory contents if available
             const lines = file?.fileContents?.split(/\r?\n/g) ?? [];